pyproject.toml
[project]
name = "weather-proxy"
version = "1.0.0"
description = "Production-ready REST API proxy for Open-Meteo weather service"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Weather Proxy Team" }]
keywords = ["weather", "api", "proxy", "flask", "redis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]

dependencies = [
    # Web framework
    "flask>=3.0.0",
    "gunicorn>=21.0.0",

    # HTTP client
    "httpx>=0.26.0",

    # Redis cache
    "redis>=5.0.0",

    # Configuration
    "python-dotenv>=1.0.0",

    # Structured logging
    "structlog>=24.0.0",

    # Resilience patterns
    "pybreaker>=1.0.0",
    "tenacity>=8.0.0",

    # Metrics (Prometheus)
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",  # For mocking httpx

    # Linting & formatting
    "ruff>=0.2.0",

    # Type checking
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
]

[project.scripts]
weather-proxy = "weather_proxy.app:create_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weather_proxy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["src/weather_proxy"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 80

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["weather_proxy"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = []
